[otbn] Adjust BN.LID/BN.SID/BN.MOVR encoding
As these use indirect addressing grd register field is moved to the
field used for grs2 in all other instructions as the given grd value
becomes a base side register to read to generate a bignum side register
index for the destination register.

BN.SID has no destination but the grs2 register field is moved to match
with all other instructions (and so the offset and inc fields match up
with BN.LID/BN.MOVR).

Fixes #3666

Signed-off-by: Greg Chadwick <gac@lowrisc.org>