[sw] Add Warnings for Implict Fallthrough
This catches a very common C pitfall: case statement bodies requiring a
`break` statement so they don't execute the next case body.

This caught a bug in `dif_i2c.c` which is now fixed.

A vendored file in CoreMark has a few errors, so I'm avoiding enforcing
this rule on that build target.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>