[sw/sca/aes_serial] Make sure to actually receive characters over UART
In lowRISC/OpenTitan#3526, a function call to inside the aes_serial app
to receive UART characters was accidentally replaced with a call to
check for available characters only. The characters were never read from
the UART Receive FIFO. As a result, the aes_serial application was not
working at all.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>