[doc] Clarify our system and software requirements
Before, we didn't explicitly list our system requirements in a
distribution-agnostic way. Instead, we provided a list of packages and a
reference distribution (Ubuntu 16.04). Users of other distributions then
had to either try, or manully cross-check if their setup works as
expected.

With this change the documentation describes better what packages are
required, and which version is required. This commit is meant to be a
clarification without changing the status quo.

Notes:
* The list of package dependencies and their versions isn't bullet-proof,
  as packages can change behavior even in-between minor
  versions (if they use semantic versioning at all). The only way to be
  sure that a distribution works is testing it. We currently do all our
  tests on Ubuntu 16.04, hence we call this the reference platform.
* OpenTitan has generally been doing a very good job at keeping
  compatibility with as many distributions and setups as
  possible, and we have been very open to feedback to improve the
  development experience for our users. We will continue to do so, but
  need user input for this: we cannot provide automated testing for all
  possible setups.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>