Update to device toolchain 20200904-1
To compile OTBN source code we need a RV32 toolchain which provides the
`as` binary, and an appropriate meson configuration to find it. This was
added in https://github.com/lowRISC/lowrisc-toolchains/pull/33, and is
part of the 20200904-1 release. Require this release to make use of this
functionality in CI.

The in-tree documentation at
https://docs.opentitan.org/doc/ug/install_instructions/#compiler-toolchain
was already updated to describe the required meson configuration.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>