[util] Add Offset Support to ROM_EXT Manifest Generator
This adds the ability to generate "offsets", which are like fields but
they don't have any size definitions. We need these for the entry point
and the interrupt vector, but I also added one so we know where the
signed area starts.

I also added support for specifying a non-default alignment, which
ensures fields and offsets are at their required alignment.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>