[memutil] Load ELF files via a staging area
This doesn't change the behaviour of LoadElfToMemories (the core of
memutil's ELF loading code), but the function now works in two stages.
The first stage loads the code into segments that are split between
the named memories (stored in staging_area_). The second stage
iterates over these memories and segments loading them into the
memory.

Of course, the point of this is that we can choose to do just the
first stage (by calling StageElf) and then programmatically get at the
segments. This will be used by a following commit to expose the
contents of the ELF over DPI to the OTBN testbench.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>