[keymgr] Revise the KMAC data interface
Adding the last beat indicator to the request interface and the error
completion in the response interface. KMAC uses last beat indicator to
initiate the hashing operation. If any error occurs during the hashing
operation, KMAC returns the garbage digest with error to indicate KeyMgr
to discard current op and re-initate.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>