[tlul] Increase Max Outstanding
Reported by Weicai in #3910

tlul socket 1n has internal variable that tracks the current outstanding
requests. It is 8 bit signal. It can have up to 255 outstandings. The
`a_source` can have up to 256 unique source IDs. When the host issues
more than 255, the internal variable overflows then the logic assumes no
outstanding.

It may not cause an issue as this issue only occurs with 1:1 connection
between the host and the device. So that the overflow still forwards the
request to the only device.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>