[otbn] Switch between model and RTL at runtime in simulations
In OTBN, we have both a RTL implementation and a model (an instruction
set simulator). Before this change, either the model or the RTL
implementation had to be selected at simulation compile time (the only
supported simulator currently is Verilator, even though there's no
blocking issue in supporting other simulators as well). After this
change the simulation can be built with both RTL and model enabled at
compile time, and then a runtime switch, a plusarg, can be used to
select between the two. It is still not possible to run both model and
RTL in parallel.

The README has been updated to show how this all works.

The main motivator for this change are CI runs: we can now use the same
Verilator-based simulation of Earl Grey to run software against both the
RTL and the model.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>