[sw, rom_exts] Add slot B linker file, and factor out common bits
LLVM RISC-V at this point does not have support for
position-independent code (PIC), which means that to produce slot A
and slot B images with the different offset from the flash start, we
must produce two executables with different absolute addresses. To do
that, we must maintain two separate linker files.

The linker script body is almost entirely the same, so most of the
functionality has been split out into the `rom_ext_common.ld`, which
is included by `rom_ext_slot_a.ld` and `rom_ext_slot_b.ld`. The main
idea is that the common body has a `.phony` section at the beginning
of flash to create an offset. The offset is defined in a
corresponding slot linker file - is 0 for slot A and half the flash
size for slot B.

Signed-off-by: Silvestrs Timofejevs <silvestrst@lowrisc.org>