[otbn/sw] Add end-to-end RSA test
Add an end-to-end test for RSA running on OTBN. Keys of 512b, 1024b,
2048b, 3072b, and 4096b can be tested with both encryption and
decryption. The encrypted data has been independently calculated and is
included as golden reference. For all keys a private key file in PEM
format is included to be able to reproduce and modify the reference
parameters. Also included is a Python script to calculate the parameters
with pycryptodome.

Note that this implementation of RSA isn't secure in any way, it's only
a hardware integration test. A comment at the beginning of the file
highlights that.

Since the test times for longer RSA keys are excessive the test contains
two constants which can be used to cut down the test time: decryption
can be disabled, and keys larger than 1024b can be left untested.

Build and run the tests with

```
 # build sw
./meson_init.sh -f
ninja -C build-out

 # build verilator simulation
fusesoc --cores-root=. run --flag=fileset_top --target=sim --setup --build lowrisc:systems:top_earlgrey_verilator

 # run the Verilator simulation, use the OTBN ISS (model)
build/lowrisc_systems_top_earlgrey_verilator_0.1/sim-verilator/Vtop_earlgrey_verilator \
  --meminit=rom,build-bin/sw/device/boot_rom/boot_rom_sim_verilator.elf  \
  --meminit=flash,build-bin/sw/device/tests/otbn_rsa_test_sim_verilator.elf \
  +UARTDPI_LOG_uart0=- +OTBN_USE_MODEL=1
```

Signed-off-by: Philipp Wagner <phw@lowrisc.org>